requirements.txt
asgiref==3.5.2
cffi==1.15.1
cryptography==37.0.4
Django==4.0.6
django-rest-knox==4.2.0
djangorestframework==3.13.1
JPype1==1.4.0
pdfkit==1.0.0
Pillow==9.2.0
psycopg2==2.9.3
pycparser==2.21
Pygments==2.13.0
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1